progress on router
